python/requirements.txt
flask
openai
qdrant_client
bs4
requests
fastembed
flask-cors
qdrant-client[fastembed]
llama-index-embeddings-openai